TextSummarizationBasedOnReviews/requirements.txt
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-auth==2.23.3
google-genai==1.38.0
pyppeteer==1.0.2